+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 <!-- My awesome changelog -->
 
+## 1.1.0 (Oct 14, 2013)
+
+IMPROVEMENTS:
+
+* Added a sample file so users can test the program without
+  having a guitar tab on their system.
+
 ## 1.0.40 (Oct 10, 2013)
 
 BUG FIXES: